+++ app/Http/Controllers/Web/ChatController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers\Web;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class ChatController extends Controller
+{
+    public function index()
+    {
+        return view('pages.web.chat.chat-room');
+    }
+}