+++ app/Livewire/Web/Chat/OnlineVipUsers.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire\Web\Chat;
+
+use Livewire\Component;
+
+class OnlineVipUsers extends Component
+{
+    public function render()
+    {
+        return view('livewire.web.chat.online-vip-users');
+    }
+}